feat:Add solutions for DP and BFS problems (Hide and Seek & Wall Breaking)
- Two problems:
DP Problem: Similar to the Baekjoon Hide and Seek problem, but with speed, acceleration, deceleration, and the constraint that the remaining distance should be less than or equal to the speed
𝑘
k before reaching the destination.

It was harder than expected, with a perceived difficulty level of Gold 3.
It is also similar to the "Fly me to the Alpha Centauri" problem.

- BFS Problem: Similar to the Baekjoon Wall Breaking problem, but in a 3D array version.

It can be solved using DP, but since
𝑛
,
𝑚
,
𝑙
≤
50
n,m,l≤50 and the movement range is limited to right, down, and upward in the z-axis, a brute force approach is feasible.
If the range increases, it should be solved using DP, considering both the number of broken walls and the directions of movement when designing the DP solution. This approach could be mixed with the "Jinwoo’s Moon Travel" problem.